pyproject.toml
[project]
name = "python-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "delta-spark>=4.0.0",
    "faker>=37.5.3",
    "fastapi>=0.116.1",
    "kafka-python>=2.2.15",
    "notebook>=7.4.3",
    "pandas>=2.3.0",
    "pyspark>=4.0.0",
    "uvicorn>=0.35.0",
]
